[Docs] Document the scrolling limitation for SSD1306
Although the OLED driver can be configured to support display width less
than 128 on the SSD1306 controller, the scrolling feature does not work
properly in that configuration, because the controller actually does not
know the display width, and uninitialized parts of the display memory
are scrolled into view.  Document this limitation for now.