Fix address calculation in calc_bounds() and calc_bounds_90()
Attempting to configure the OLED driver for a 64×48 display uncovered
some errors in the address calculations.  Other displays apparently
always used blocks that occupied the whole physical height with 90°/270°
rotation, but the 64×48 display requires using smaller blocks
(specifically, 24 blocks of 16 bytes, so that both height and width
could be divided into whole blocks), therefore the top and bottom page
numbers in the address need to be calculated properly.

There was also a minor error in the end page number calculation in
calc_bounds(), but it did not seem to affect anything (maybe that value
is not actually used by the controller unless blocks actually span
multiple pages).