Implement 90°/270° rotation for SH1106 and SH1107
Because the SH1106 and SH1107 chips do not support the horizontal
addressing mode, it is not possible to transfer the whole block at once
when using 90°/270° rotation.  Use separate transfers with address setup
for every page-wide strip instead; this is slower than doing a single
transfer, but still better than no rotation support at all.